pkg/reporting/calculator.go
package reporting

import (
	"context"
	"fmt"
	"time"

	"github.com/johnayoung/finlib/pkg/account"
	"github.com/johnayoung/finlib/pkg/money"
	"github.com/shopspring/decimal"
)

// defaultReportCalculator implements the ReportCalculator interface
type defaultReportCalculator struct {
	accountStore account.Repository
}

// NewReportCalculator creates a new instance of the report calculator
func NewReportCalculator(accountStore account.Repository) ReportCalculator {
	return &defaultReportCalculator{
		accountStore: accountStore,
	}
}

// CalculateBalance computes account balances for reporting
func (c *defaultReportCalculator) CalculateBalance(ctx context.Context, accountID string, period ReportPeriod) (money.Money, error) {
	// Get the account
	var acc account.Account
	if err := c.accountStore.Read(ctx, accountID, &acc); err != nil {
		return money.Money{}, fmt.Errorf("error reading account: %w", err)
	}

	// Get all transactions for the account within the period
	transactions, err := c.getTransactions(ctx, accountID, period)
	if err != nil {
		return money.Money{}, fmt.Errorf("error getting transactions: %w", err)
	}

	// Calculate the balance based on transactions
	balance := c.calculateBalanceFromTransactions(transactions, acc.Type)

	return balance, nil
}

// CalculateChanges computes changes over a period
func (c *defaultReportCalculator) CalculateChanges(ctx context.Context, accountID string, period ReportPeriod) (*BalanceChange, error) {
	// Get opening balance
	openingBalance, err := c.getBalanceAtTime(ctx, accountID, period.Start)
	if err != nil {
		return nil, fmt.Errorf("error getting opening balance: %w", err)
	}

	// Get closing balance
	closingBalance, err := c.getBalanceAtTime(ctx, accountID, period.End)
	if err != nil {
		return nil, fmt.Errorf("error getting closing balance: %w", err)
	}

	// Get movements during the period
	movements, err := c.getMovements(ctx, accountID, period)
	if err != nil {
		return nil, fmt.Errorf("error getting movements: %w", err)
	}

	// Calculate net change
	netChange := money.Money{
		Amount:   closingBalance.Amount.Sub(openingBalance.Amount),
		Currency: closingBalance.Currency,
	}

	return &BalanceChange{
		OpeningBalance: openingBalance,
		ClosingBalance: closingBalance,
		NetChange:      netChange,
		Movements:      movements,
	}, nil
}

// CalculateRatio computes financial ratios
func (c *defaultReportCalculator) CalculateRatio(ctx context.Context, ratio RatioDefinition, period ReportPeriod) (decimal.Decimal, error) {
	// Calculate numerator
	numerator, err := c.calculateValue(ctx, ratio.Numerator, period)
	if err != nil {
		return decimal.Zero, fmt.Errorf("error calculating numerator: %w", err)
	}

	// Calculate denominator
	denominator, err := c.calculateValue(ctx, ratio.Denominator, period)
	if err != nil {
		return decimal.Zero, fmt.Errorf("error calculating denominator: %w", err)
	}

	// Check for division by zero
	if denominator.IsZero() {
		return decimal.Zero, fmt.Errorf("division by zero: denominator is zero")
	}

	// Calculate ratio with proper scaling
	result := numerator.Div(denominator)
	if ratio.Scale > 0 {
		result = result.Round(ratio.Scale)
	}

	return result, nil
}

// Helper functions

func (c *defaultReportCalculator) getTransactions(ctx context.Context, accountID string, period ReportPeriod) ([]Transaction, error) {
	// This would be implemented to fetch transactions from a transaction store
	// For now, return a not implemented error
	return nil, fmt.Errorf("getTransactions not implemented")
}

func (c *defaultReportCalculator) calculateBalanceFromTransactions(transactions []Transaction, accountType account.AccountType) money.Money {
	// This would implement the actual balance calculation logic
	// For now, return a zero balance
	return money.Money{
		Amount:   decimal.Zero,
		Currency: "USD", // Default currency, should be configurable
	}
}

func (c *defaultReportCalculator) getBalanceAtTime(ctx context.Context, accountID string, at time.Time) (money.Money, error) {
	// This would calculate the balance at a specific point in time
	// For now, return a not implemented error
	return money.Money{}, fmt.Errorf("getBalanceAtTime not implemented")
}

func (c *defaultReportCalculator) getMovements(ctx context.Context, accountID string, period ReportPeriod) ([]BalanceMovement, error) {
	// This would get all balance movements in a period
	// For now, return a not implemented error
	return nil, fmt.Errorf("getMovements not implemented")
}

func (c *defaultReportCalculator) calculateValue(ctx context.Context, calc Calculation, period ReportPeriod) (decimal.Decimal, error) {
	// This would implement the calculation logic based on the calculation type
	// For now, return a not implemented error
	return decimal.Zero, fmt.Errorf("calculateValue not implemented")
}

// Transaction represents a financial transaction
// This should be moved to a separate package when implementing the full transaction system
type Transaction struct {
	ID          string
	AccountID   string
	Amount      money.Money
	Type        string
	Date        time.Time
	Description string
	Metadata    map[string]interface{}
}
