pkg/reporting/calculator_test.go
package reporting

import (
	"context"
	"testing"
	"time"

	"github.com/johnayoung/finlib/pkg/account"
	"github.com/johnayoung/finlib/pkg/money"
	"github.com/shopspring/decimal"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/mock"
)

// Mock account repository
type mockAccountRepository struct {
	mock.Mock
}

func (m *mockAccountRepository) Create(ctx context.Context, entity interface{}) error {
	args := m.Called(ctx, entity)
	return args.Error(0)
}

func (m *mockAccountRepository) Read(ctx context.Context, id string, entity interface{}) error {
	args := m.Called(ctx, id, entity)
	if acc, ok := args.Get(0).(*account.Account); ok && acc != nil {
		*(entity.(*account.Account)) = *acc
	}
	return args.Error(1)
}

func (m *mockAccountRepository) Update(ctx context.Context, entity interface{}) error {
	args := m.Called(ctx, entity)
	return args.Error(0)
}

func (m *mockAccountRepository) Delete(ctx context.Context, id string) error {
	args := m.Called(ctx, id)
	return args.Error(0)
}

func (m *mockAccountRepository) Query(ctx context.Context, query interface{}, results interface{}) error {
	args := m.Called(ctx, query, results)
	return args.Error(0)
}

func TestNewReportCalculator(t *testing.T) {
	accountStore := &mockAccountRepository{}
	calculator := NewReportCalculator(accountStore)
	assert.NotNil(t, calculator)
}

func TestCalculateBalance(t *testing.T) {
	// Setup
	ctx := context.Background()
	accountStore := &mockAccountRepository{}
	calculator := NewReportCalculator(accountStore)

	testTime := time.Date(2024, 12, 24, 10, 0, 0, 0, time.UTC)
	testAccount := &account.Account{
		ID:      "ACC001",
		Type:    account.Asset,
		Balance: &money.Money{Amount: decimal.NewFromInt(1000), Currency: "USD"},
	}

	period := ReportPeriod{
		Start: testTime.AddDate(0, -1, 0),
		End:   testTime,
	}

	// Setup mock expectations
	accountStore.On("Read", ctx, "ACC001", mock.AnythingOfType("*account.Account")).
		Run(func(args mock.Arguments) {
			acc := args.Get(2).(*account.Account)
			*acc = *testAccount
		}).
		Return(testAccount, nil)

	// Execute test
	_, err := calculator.CalculateBalance(ctx, "ACC001", period)

	// Since getTransactions is not implemented, we expect an error
	assert.Error(t, err)
	assert.Contains(t, err.Error(), "getTransactions not implemented")
}

func TestCalculateChanges(t *testing.T) {
	// Setup
	ctx := context.Background()
	accountStore := &mockAccountRepository{}
	calculator := NewReportCalculator(accountStore)

	testTime := time.Date(2024, 12, 24, 10, 0, 0, 0, time.UTC)
	period := ReportPeriod{
		Start: testTime.AddDate(0, -1, 0),
		End:   testTime,
	}

	// Execute test
	changes, err := calculator.CalculateChanges(ctx, "ACC001", period)

	// Since getBalanceAtTime is not implemented, we expect an error
	assert.Error(t, err)
	assert.Nil(t, changes)
	assert.Contains(t, err.Error(), "getBalanceAtTime not implemented")
}

func TestCalculateRatio(t *testing.T) {
	// Setup
	ctx := context.Background()
	accountStore := &mockAccountRepository{}
	calculator := NewReportCalculator(accountStore)

	testTime := time.Date(2024, 12, 24, 10, 0, 0, 0, time.UTC)
	period := ReportPeriod{
		Start: testTime.AddDate(0, -1, 0),
		End:   testTime,
	}

	ratio := RatioDefinition{
		ID:          "CURRENT_RATIO",
		Name:        "Current Ratio",
		Description: "Current Assets / Current Liabilities",
		Scale:       2,
		Numerator: Calculation{
			ID:   "CURRENT_ASSETS",
			Type: "BALANCE",
			AccountSelector: AccountSelector{
				Types: []account.AccountType{account.Asset},
			},
		},
		Denominator: Calculation{
			ID:   "CURRENT_LIABILITIES",
			Type: "BALANCE",
			AccountSelector: AccountSelector{
				Types: []account.AccountType{account.Liability},
			},
		},
	}

	// Execute test
	result, err := calculator.CalculateRatio(ctx, ratio, period)

	// Since calculateValue is not implemented, we expect an error
	assert.Error(t, err)
	assert.True(t, result.IsZero())
	assert.Contains(t, err.Error(), "calculateValue not implemented")
}

func TestCalculateBalanceFromTransactions(t *testing.T) {
	calculator := &defaultReportCalculator{}

	transactions := []Transaction{
		{
			ID:        "TXN001",
			AccountID: "ACC001",
			Amount:    money.Money{Amount: decimal.NewFromInt(100), Currency: "USD"},
			Type:      "CREDIT",
			Date:      time.Now(),
		},
		{
			ID:        "TXN002",
			AccountID: "ACC001",
			Amount:    money.Money{Amount: decimal.NewFromInt(50), Currency: "USD"},
			Type:      "DEBIT",
			Date:      time.Now(),
		},
	}

	balance := calculator.calculateBalanceFromTransactions(transactions, account.Asset)
	assert.Equal(t, "USD", balance.Currency)
	assert.True(t, balance.Amount.IsZero()) // Currently returns zero as it's not implemented
}
